Add files via upload
the naive
Bayes probabilistic classifier is often used in text classification applications and experiments
because of its simplicity and effectiveness which basic idea is to use the joint probabilities of words
and categories to estimate the probabilities of categories given a document. They are simple
probabilistic classifiers based on Bayes' theorem with strong independence assumed between
features given a class, while “the independence assumptions on which naive Bayes classifiers are
based almost never hold for natural data sets”. (Lewis 11). Softmax regression as a deep learning
algorithm and neural network models the relationship between features and the probabilities of
belonging to each label using the softmax function. Since deep learning algorithms allegedly have
better performance for text classification tasks such as sentiment analysis (Sunarya et al. 77), this
paper aims to compare the accuracy of naive Bayes and softmax regression from this perspective.